+++ src/sqlfmt/exception.py
@@ -6,8 +6,9 @@ class SqlfmtError(ValueError):
     """
 
     def __str__(self) -> str:
+        intro = "sqlfmt encountered an error: "
         message = super().__str__()
-        return f"sqlfmt encountered an error: {message}"
+        return f"{intro}{message}"
 
 
 class SqlfmtParsingError(SqlfmtError):